Anime2.0/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.0
openai==1.3.5
replicate==0.22.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.1
pillow>=10.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
moviepy==1.0.3